Перенес getWatchFilms в main-navigation.js
