+++ src/mock/main-navigation.js
@@ -1,4 +1,23 @@
-export const generateItems = (item) => {
+/**
+ * @description Показываем общее число просмотренных/избранных фильмов
+ * @param films - массив фильмов
+ * @returns {*} объект чисел на каждый из категории фильтра
+ */
+
+const getWatchFilms = (films) => films.reduce((stats, film) => {
+  if (film.isWatchlist) {
+    stats.watchlist += 1;
+  }
+  if (film.isHistory) {
+    stats.history += 1;
+  }
+  if (film.isFavorites) {
+    stats.favorites += 1;
+  }
+  return stats;
+}, {watchlist: 0, history: 0, favorites: 0});
+
+const generateItems = (item) => {
   return [{
     name: `All`,
   }, {
@@ -12,3 +31,5 @@ export const generateItems = (item) => {
     count: item.favorites,
   }];
 };
+
+export {generateItems, getWatchFilms};
